Revert "Kernel/Net: Rename IPv4-specific files and structs"
This reverts commit
ad73adef5d2f8e198553441657c544b278605e8a
partially. All base file names and structs are
restored, though the new directory is kept. The IP
directory will soon contain IPv6-related files, so
the directory should be named "IP" instead.
