+++ Kernel/Net/NetworkAdapter.h
@@ -19,9 +19,9 @@
 #include <Kernel/Library/LockWeakable.h>
 #include <Kernel/Library/UserOrKernelBuffer.h>
 #include <Kernel/Net/EthernetFrameHeader.h>
-#include <Kernel/Net/ICMP.h>
-#include <Kernel/Net/IPv4/ARP.h>
-#include <Kernel/Net/IPv4/IPv4.h>
+#include <Kernel/Net/IP/ARP.h>
+#include <Kernel/Net/IP/ICMP.h>
+#include <Kernel/Net/IP/IPv4.h>
 
 namespace Kernel {
