+++ Kernel/Net/TCPSocket.h
@@ -14,7 +14,8 @@
 #include <AK/Time.h>
 #include <Kernel/Library/LockWeakPtr.h>
 #include <Kernel/Locking/MutexProtected.h>
-#include <Kernel/Net/IPv4/Socket.h>
+#include <Kernel/Net/IP/IPv4Socket.h>
+#include <Kernel/Net/Routing.h>
 #include <Kernel/Time/TimerQueue.h>
 
 namespace Kernel {
