+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -696,7 +696,7 @@ executable("Kernel_bin") {
     "Memory/SharedInodeVMObject.cpp",
     "Memory/VMObject.cpp",
     "Memory/VirtualRange.cpp",
-    "Net/IPv4/Socket.cpp",
+    "Net/IP/IPv4Socket.cpp",
     "Net/Intel/E1000ENetworkAdapter.cpp",
     "Net/Intel/E1000NetworkAdapter.cpp",
     "Net/LocalSocket.cpp",
